# Hand SAT optimization for Data.ByteString.unfoldrN