This follows the approach described in #350 to optimize `unfoldrN` to float out `p` and use `pokeByteOff`