@Bodigrim 2021-01-30
Cool! Could you please add a benchmark to `bench/BenchAll.hs`, so that we can estimate performance gains?

And if you throw `{-# INLINE createAndTrim' #-}` atop, we'll be able to resolve #128 as well.